\name{paniciR-package}
\alias{paniciR-package}
\alias{paniciR}
\docType{package}
\title{
\packageTitle{paniciR}
}
\description{
\packageDescription{paniciR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{paniciR}
\packageIndices{paniciR}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{paniciR}

Maintainer: \packageMaintainer{paniciR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
